requirements.txt
streamlit>=1.20.0
pdfplumber>=0.7.6
pypdf>=3.15.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4         
transformers>=4.35.0
torch>=2.0.0            
tqdm
python-multipart
